EX-109-add-api-edit-api-to-oms: OMS order service improvements: fix Flyway schema (orders table), replace double with BigDecimal, add balance validation

+++ oms/src/main/java/com/exchange/oms/client/matchingengine/createOrderRequest.java
@@ -5,8 +5,10 @@ import com.exchange.oms.domain.OrderType;
 import com.exchange.oms.domain.TradePair;
 import com.exchange.oms.domain.TradeSide;
 
+import java.math.BigDecimal;
+
 public record createOrderRequest(long orderId, long userId,
                                  TradePair tradePair, OrderType orderType,
                                  TradeSide tradeSide,
-                                 double quantity, double price) {
+                                 BigDecimal quantity, BigDecimal price) {
 }
