+++ oms/src/main/java/com/exchange/oms/domain/Order.java
@@ -8,6 +8,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -20,19 +21,29 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long userId ;
+
+    private Long userId;
+
     @Enumerated(EnumType.STRING)
     private OrderType orderType;
+
     @Enumerated(EnumType.STRING)
     private TradeSide tradeSide;
+
     @Enumerated(EnumType.STRING)
     private TradePair tradePair;
+
     @Enumerated(EnumType.STRING)
     private OrderStatus status = OrderStatus.NEW;
-    private double quantity;
-    private double price;
+
+    private BigDecimal quantity;
+
+    private BigDecimal price;
+
     private MatchEventStatus matchEngineStatus;
+
     private LocalDateTime createdAt;
+
     private LocalDateTime updatedAt;
 
 }
